+++ app/src/main/java/com/actito/go/ActitoApp.kt
@@ -5,10 +5,13 @@ import android.os.Build
 import androidx.hilt.work.HiltWorkerFactory
 import androidx.work.Configuration
 import com.actito.Actito
+import com.actito.go.core.configure
 import com.actito.go.live_activities.LiveActivitiesController
 import com.actito.go.storage.preferences.ActitoSharedPreferences
 import com.actito.push.ktx.push
 import com.google.android.material.color.DynamicColors
+import com.google.firebase.Firebase
+import com.google.firebase.crashlytics.crashlytics
 import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
 import javax.inject.Inject
@@ -42,9 +45,14 @@ class ActitoApp : Application(), Configuration.Provider {
         // Configure Actito if there is a stored configuration set.
         val configuration = preferences.appConfiguration
         if (configuration != null) {
-            Actito.configure(this, configuration.applicationKey, configuration.applicationSecret)
+            configure(this, configuration)
+
+            Actito.device().currentDevice?.let { device ->
+                Firebase.crashlytics.setUserId("device_${device.id}")
+            }
         }
 
+        Actito.intentReceiver = CustomIntentReceiver::class.java
         Actito.push().intentReceiver = PushReceiver::class.java
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {