+++ app/src/main/java/com/actito/go/ui/MainViewModel.kt
@@ -7,14 +7,14 @@ import androidx.lifecycle.viewModelScope
 import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkManager
 import com.actito.Actito
+import com.actito.go.core.configure
 import com.actito.go.core.createDynamicShortcuts
 import com.actito.go.core.loadRemoteConfig
 import com.actito.go.models.AppConfiguration
-import com.actito.go.network.push.PushService
+import com.actito.go.network.push.PushServiceFactory
 import com.actito.go.storage.preferences.ActitoSharedPreferences
 import com.actito.go.workers.UpdateProductsWorker
 import com.actito.iam.ktx.inAppMessaging
-import com.actito.ktx.device
 import com.actito.models.ActitoApplication
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
@@ -34,7 +34,7 @@ import javax.inject.Inject
 class MainViewModel @Inject constructor(
     @param:ApplicationContext private val context: Context,
     private val preferences: ActitoSharedPreferences,
-    private val pushService: PushService,
+    private val pushServiceFactory: PushServiceFactory,
     private val workManager: WorkManager,
 ) : ViewModel(), Actito.Listener {
 
@@ -70,14 +70,17 @@ class MainViewModel @Inject constructor(
     }
 
 
-    suspend fun configure(code: String): ConfigurationResult = withContext(Dispatchers.IO) {
+    suspend fun configure(code: String, environment: AppConfiguration.Environment): ConfigurationResult = withContext(Dispatchers.IO) {
         if (hasConfiguration) return@withContext ConfigurationResult.ALREADY_CONFIGURED
+
+        val pushService = pushServiceFactory.createService(environment.baseUrl)
 
         val configuration = pushService.getConfiguration(code).let {
             AppConfiguration(
                 applicationKey = it.demo.applicationKey,
                 applicationSecret = it.demo.applicationSecret,
                 loyaltyProgramId = it.demo.loyaltyProgram,
+                environment = environment
             )
         }
 
@@ -95,7 +98,7 @@ class MainViewModel @Inject constructor(
             ?: throw IllegalStateException("Cannot launch Actito before the application has been configured.")
 
         if (!Actito.isConfigured) {
-            Actito.configure(context, configuration.applicationKey, configuration.applicationSecret)
+            configure(context, configuration)
         }
 
         // Let's get started! 🚀
@@ -137,6 +140,12 @@ class MainViewModel @Inject constructor(
 
             navigationChannel.trySend(NavigationOption.MAIN)
         }
+    }
+
+    override fun onUnlaunched() {
+        preferences.resetPreferences()
+        navigationChannel.trySend(NavigationOption.SPLASH)
+        launch()
     }
 
     // endregion