+++ app/src/main/java/com/actito/go/ui/home/HomeViewModel.kt
@@ -21,7 +21,6 @@ import com.actito.go.models.Product
 import com.actito.go.storage.db.ActitoDatabase
 import com.actito.go.storage.db.mappers.toModel
 import com.actito.go.storage.preferences.ActitoSharedPreferences
-import com.actito.ktx.events
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOn
@@ -60,6 +59,14 @@ class HomeViewModel @Inject constructor(
                 .collect { products ->
                     _products.postValue(products.map { it.toModel() })
                 }
+        }
+    }
+
+    fun changeLocationUpdates(enabled: Boolean) {
+        if (enabled) {
+            Actito.geo().enableLocationUpdates()
+        } else {
+            Actito.geo().disableLocationUpdates()
         }
     }