+++ app/src/main/java/com/actito/go/ui/main/MainFragment.kt
@@ -7,7 +7,6 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
 import androidx.core.view.isGone
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
@@ -52,16 +51,11 @@ class MainFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         requireActivity().window.navigationBarColor = SurfaceColors.SURFACE_2.getColor(requireActivity())
 
-        ViewCompat.setOnApplyWindowInsetsListener(binding.bottomNavigation) { view, insets ->
-            view.setPadding(0, 0, 0, 0)
-            insets
-        }
-        
         // Only shows the cart tab when enabled in the remote config.
-        binding.bottomNavigation.menu.findItem(R.id.cart_fragment).isVisible = sharedPreferences.hasStoreEnabled
+        binding.bottomNavigation.menu.findItem(R.id.cart_nested_graph).isVisible = sharedPreferences.hasStoreEnabled
 
         val appBarConfiguration = AppBarConfiguration(
-            setOf(R.id.home_fragment, R.id.cart_fragment, R.id.settings_fragment)
+            setOf(R.id.home_nested_graph, R.id.cart_nested_graph, R.id.settings_nested_graph)
         )
 
         (activity as AppCompatActivity).setSupportActionBar(binding.toolbar)