+++ app/src/main/java/com/actito/go/ui/product_details/ProductDetailsViewModel.kt
@@ -15,7 +15,6 @@ import com.actito.go.models.Product
 import com.actito.go.storage.db.ActitoDatabase
 import com.actito.go.storage.db.entities.CartEntryEntity
 import com.actito.go.storage.db.mappers.toModel
-import com.actito.ktx.events
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch