+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 2.1.0
+
+- Fix card icons tint in dark mode
+- Improve small screens UI
+- Add UI flow to enable beacon scanning when location capabilities are missing
+- Improve short links navigation
+- Permanently remove all data when deleting account
+- Update dependencies
+
 ## 2.0.0
 
 - Migrate to Actito SDK 5.0.0-beta.1